requirements.txt
appnope==0.1.3
asgiref==3.7.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.0
colorama==0.4.6
cryptography==41.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-bootstrap4==23.1
django-classy-tags==4.0.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-markdownx==4.0.2
django-summernote==0.8.20.0
django-taggit==4.0.0
django-taggit-labels==0.7.0
django-taggit-templatetags2==1.6.1
executing==1.2.0
filelock==3.12.0
idna==3.3
ipython==8.13.2
jedi==0.18.2
Markdown==3.4.3
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyproject_api==1.5.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tox==4.6.0
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==1.26.10
virtualenv==20.23.0
wcwidth==0.2.6
webencodings==0.5.1
